- photographerId changed to be Long type

+++ app/src/main/java/com/example/j_pexels_app/data/remote/dto/PhotoDetailsResponse.kt
@@ -11,7 +11,7 @@ data class PhotoDetailsResponse(
     @Json(name = "photographer_url")
     val photographerUrl: String = "",
     @Json(name = "photographer_id")
-    val photographerId: Int = 0,
+    val photographerId: Long = 0L,
     @Json(name = "avg_color")
     val avgColor: String = "",
     @Json(name = "src")
